third_party/WebKit/LayoutTests/fast/events/pointerevents/mouse-pointer-capture-transition-events-expected.txt

Verifies that trasition events are correctly fired in different pointer capture scenarios for mouse.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


    ==== 'none' is capturing and then 'blue' will capture; will move pointer to 'blue' ====
 **** Move mouse to green and press *****
green received pointerover
grey received pointerenter
green received pointerenter
green received mouseover
grey received mouseenter
green received mouseenter
green received pointermove
green received mousemove
 **** Set initial capturing and jiggle mouse in green *****
green received pointermove
green received mousemove
--- Set pointercapture for blue ---
 **** Move to blue box *****
green received pointerout
green received pointerleave
blue received pointerover
blue received pointerenter
blue received gotpointercapture
green received mouseout
green received mouseleave
blue received mouseover
blue received mouseenter
blue received pointermove
blue received mousemove
 ****  Move out of grey, then release *****
blue received pointermove
blue received mousemove
blue received lostpointercapture
blue received pointerout
blue received pointerleave
grey received pointerleave
blue received mouseout
blue received mouseleave
grey received mouseleave

    ==== 'none' is capturing and then 'green' will capture; will move pointer to 'blue' ====
 **** Move mouse to green and press *****
green received pointerover
grey received pointerenter
green received pointerenter
green received mouseover
grey received mouseenter
green received mouseenter
green received pointermove
green received mousemove
 **** Set initial capturing and jiggle mouse in green *****
green received pointermove
green received mousemove
--- Set pointercapture for green ---
 **** Move to blue box *****
green received gotpointercapture
green received pointermove
green received mousemove
 ****  Move out of grey, then release *****
green received pointermove
green received mousemove
green received lostpointercapture
green received pointerout
green received pointerleave
grey received pointerleave
green received mouseout
green received mouseleave
grey received mouseleave

    ==== 'blue' is capturing and then 'none' will capture; will move pointer to 'blue' ====
 **** Move mouse to green and press *****
green received pointerover
grey received pointerenter
green received pointerenter
green received mouseover
grey received mouseenter
green received mouseenter
green received pointermove
green received mousemove
 **** Set initial capturing and jiggle mouse in green *****
--- Set pointercapture for blue ---
green received pointerout
green received pointerleave
blue received pointerover
blue received pointerenter
blue received gotpointercapture
green received mouseout
green received mouseleave
blue received mouseover
blue received mouseenter
blue received pointermove
blue received mousemove
--- Release pointercapture for blue ---
 **** Move to blue box *****
blue received lostpointercapture
blue received pointermove
blue received mousemove
 ****  Move out of grey, then release *****
blue received pointerout
blue received pointerleave
grey received pointerleave
blue received mouseout
blue received mouseleave
grey received mouseleave

    ==== 'green' is capturing and then 'none' will capture; will move pointer to 'blue' ====
 **** Move mouse to green and press *****
green received pointerover
grey received pointerenter
green received pointerenter
green received mouseover
grey received mouseenter
green received mouseenter
green received pointermove
green received mousemove
 **** Set initial capturing and jiggle mouse in green *****
--- Set pointercapture for green ---
green received gotpointercapture
green received pointermove
green received mousemove
--- Release pointercapture for green ---
 **** Move to blue box *****
green received lostpointercapture
green received pointerout
green received pointerleave
blue received pointerover
blue received pointerenter
green received mouseout
green received mouseleave
blue received mouseover
blue received mouseenter
blue received pointermove
blue received mousemove
 ****  Move out of grey, then release *****
blue received pointerout
blue received pointerleave
grey received pointerleave
blue received mouseout
blue received mouseleave
grey received mouseleave

    ==== 'blue' is capturing and then 'green' will capture; will move pointer to 'blue' ====
 **** Move mouse to green and press *****
green received pointerover
grey received pointerenter
green received pointerenter
green received mouseover
grey received mouseenter
green received mouseenter
green received pointermove
green received mousemove
 **** Set initial capturing and jiggle mouse in green *****
--- Set pointercapture for blue ---
green received pointerout
green received pointerleave
blue received pointerover
blue received pointerenter
blue received gotpointercapture
green received mouseout
green received mouseleave
blue received mouseover
blue received mouseenter
blue received pointermove
blue received mousemove
--- Set pointercapture for green ---
 **** Move to blue box *****
blue received lostpointercapture
blue received pointerout
blue received pointerleave
green received pointerover
green received pointerenter
green received gotpointercapture
blue received mouseout
blue received mouseleave
green received mouseover
green received mouseenter
green received pointermove
green received mousemove
 ****  Move out of grey, then release *****
green received pointermove
green received mousemove
green received lostpointercapture
green received pointerout
green received pointerleave
grey received pointerleave
green received mouseout
green received mouseleave
grey received mouseleave

    ==== 'green' is capturing and then 'blue' will capture; will move pointer to 'blue' ====
 **** Move mouse to green and press *****
green received pointerover
grey received pointerenter
green received pointerenter
green received mouseover
grey received mouseenter
green received mouseenter
green received pointermove
green received mousemove
 **** Set initial capturing and jiggle mouse in green *****
--- Set pointercapture for green ---
green received gotpointercapture
green received pointermove
green received mousemove
--- Set pointercapture for blue ---
 **** Move to blue box *****
green received lostpointercapture
green received pointerout
green received pointerleave
blue received pointerover
blue received pointerenter
blue received gotpointercapture
green received mouseout
green received mouseleave
blue received mouseover
blue received mouseenter
blue received pointermove
blue received mousemove
 ****  Move out of grey, then release *****
blue received pointermove
blue received mousemove
blue received lostpointercapture
blue received pointerout
blue received pointerleave
grey received pointerleave
blue received mouseout
blue received mouseleave
grey received mouseleave

    ==== 'none' is capturing and then 'blue' will capture ====
 **** Move mouse to green and press *****
green received pointerover
grey received pointerenter
green received pointerenter
green received mouseover
grey received mouseenter
green received mouseenter
green received pointermove
green received mousemove
 **** Set initial capturing and jiggle mouse in green *****
green received pointermove
green received mousemove
--- Set pointercapture for blue ---
 **** Jiggle mouse in green *****
green received pointerout
green received pointerleave
blue received pointerover
blue received pointerenter
blue received gotpointercapture
green received mouseout
green received mouseleave
blue received mouseover
blue received mouseenter
blue received pointermove
blue received mousemove
 ****  Move out of grey, then release *****
blue received pointermove
blue received mousemove
blue received lostpointercapture
blue received pointerout
blue received pointerleave
grey received pointerleave
blue received mouseout
blue received mouseleave
grey received mouseleave

    ==== 'none' is capturing and then 'green' will capture ====
 **** Move mouse to green and press *****
green received pointerover
grey received pointerenter
green received pointerenter
green received mouseover
grey received mouseenter
green received mouseenter
green received pointermove
green received mousemove
 **** Set initial capturing and jiggle mouse in green *****
green received pointermove
green received mousemove
--- Set pointercapture for green ---
 **** Jiggle mouse in green *****
green received gotpointercapture
green received pointermove
green received mousemove
 ****  Move out of grey, then release *****
green received pointermove
green received mousemove
green received lostpointercapture
green received pointerout
green received pointerleave
grey received pointerleave
green received mouseout
green received mouseleave
grey received mouseleave

    ==== 'blue' is capturing and then 'none' will capture ====
 **** Move mouse to green and press *****
green received pointerover
grey received pointerenter
green received pointerenter
green received mouseover
grey received mouseenter
green received mouseenter
green received pointermove
green received mousemove
 **** Set initial capturing and jiggle mouse in green *****
--- Set pointercapture for blue ---
green received pointerout
green received pointerleave
blue received pointerover
blue received pointerenter
blue received gotpointercapture
green received mouseout
green received mouseleave
blue received mouseover
blue received mouseenter
blue received pointermove
blue received mousemove
--- Release pointercapture for blue ---
 **** Jiggle mouse in green *****
blue received lostpointercapture
blue received pointerout
blue received pointerleave
green received pointerover
green received pointerenter
blue received mouseout
blue received mouseleave
green received mouseover
green received mouseenter
green received pointermove
green received mousemove
 ****  Move out of grey, then release *****
green received pointerout
green received pointerleave
grey received pointerleave
green received mouseout
green received mouseleave
grey received mouseleave

    ==== 'green' is capturing and then 'none' will capture ====
 **** Move mouse to green and press *****
green received pointerover
grey received pointerenter
green received pointerenter
green received mouseover
grey received mouseenter
green received mouseenter
green received pointermove
green received mousemove
 **** Set initial capturing and jiggle mouse in green *****
--- Set pointercapture for green ---
green received gotpointercapture
green received pointermove
green received mousemove
--- Release pointercapture for green ---
 **** Jiggle mouse in green *****
green received lostpointercapture
green received pointermove
green received mousemove
 ****  Move out of grey, then release *****
green received pointerout
green received pointerleave
grey received pointerleave
green received mouseout
green received mouseleave
grey received mouseleave

    ==== 'blue' is capturing and then 'green' will capture ====
 **** Move mouse to green and press *****
green received pointerover
grey received pointerenter
green received pointerenter
green received mouseover
grey received mouseenter
green received mouseenter
green received pointermove
green received mousemove
 **** Set initial capturing and jiggle mouse in green *****
--- Set pointercapture for blue ---
green received pointerout
green received pointerleave
blue received pointerover
blue received pointerenter
blue received gotpointercapture
green received mouseout
green received mouseleave
blue received mouseover
blue received mouseenter
blue received pointermove
blue received mousemove
--- Set pointercapture for green ---
 **** Jiggle mouse in green *****
blue received lostpointercapture
blue received pointerout
blue received pointerleave
green received pointerover
green received pointerenter
green received gotpointercapture
blue received mouseout
blue received mouseleave
green received mouseover
green received mouseenter
green received pointermove
green received mousemove
 ****  Move out of grey, then release *****
green received pointermove
green received mousemove
green received lostpointercapture
green received pointerout
green received pointerleave
grey received pointerleave
green received mouseout
green received mouseleave
grey received mouseleave

    ==== 'green' is capturing and then 'blue' will capture ====
 **** Move mouse to green and press *****
green received pointerover
grey received pointerenter
green received pointerenter
green received mouseover
grey received mouseenter
green received mouseenter
green received pointermove
green received mousemove
 **** Set initial capturing and jiggle mouse in green *****
--- Set pointercapture for green ---
green received gotpointercapture
green received pointermove
green received mousemove
--- Set pointercapture for blue ---
 **** Jiggle mouse in green *****
green received lostpointercapture
green received pointerout
green received pointerleave
blue received pointerover
blue received pointerenter
blue received gotpointercapture
green received mouseout
green received mouseleave
blue received mouseover
blue received mouseenter
blue received pointermove
blue received mousemove
 ****  Move out of grey, then release *****
blue received pointermove
blue received mousemove
blue received lostpointercapture
blue received pointerout
blue received pointerleave
grey received pointerleave
blue received mouseout
blue received mouseleave
grey received mouseleave

PASS successfullyParsed is true

TEST COMPLETE
